Update script to create obfuscated UFC
