+++ package.json
@@ -25,7 +25,7 @@
     "typecheck": "tsc",
     "test": "yarn test:unit",
     "test:unit": "NODE_ENV=test jest '.*\\.spec\\.ts'",
-    "obfuscate-mock-rac": "ts-node test/writeObfuscatedMockRac"
+    "obfuscate-mock-ufc": "ts-node test/writeObfuscatedMockUFC"
   },
   "jsdelivr": "dist/eppo-sdk.js",
   "repository": {
